+++ Sources/Sentry/SentryDsn.m
@@ -9,12 +9,15 @@ NS_ASSUME_NONNULL_BEGIN
 
 @end
 
-@implementation SentryDsn
+@implementation SentryDsn {
+    NSURL *_storeEndpoint;
+    NSURL *_envelopeEndpoint;
+}
 
 - (_Nullable instancetype)initWithString:(NSString *)dsnString didFailWithError:(NSError *_Nullable *_Nullable)error {
     self = [super init];
     if (self) {
-        self.url = [self convertDsnString:dsnString didFailWithError:error];
+        _url = [self convertDsnString:dsnString didFailWithError:error];
         if (nil != error && nil != *error) {
             return nil;
         }
@@ -31,6 +34,50 @@ NS_ASSUME_NONNULL_BEGIN
         [output appendFormat:@"%02x", digest[i]];
     }
     return output;
+}
+
+- (NSURL *)getStoreEndpoint {
+    if (nil == _storeEndpoint) {
+        @synchronized (self) {
+            if (nil == _storeEndpoint) {
+                _storeEndpoint = [[self getBaseEndpoint] URLByAppendingPathComponent:@"store/"];
+            }
+        }
+    }
+    return _storeEndpoint;
+}
+
+- (NSURL *)getEnvelopeEndpoint {
+    if (nil == _envelopeEndpoint) {
+        @synchronized (self) {
+            if (nil == _envelopeEndpoint) {
+                _envelopeEndpoint = [[self getBaseEndpoint] URLByAppendingPathComponent:@"envelope/"];
+            }
+        }
+    }
+    return _envelopeEndpoint;
+}
+
+- (NSURL *)getBaseEndpoint {
+    NSURL *url = self.url;
+    NSString *projectId = url.lastPathComponent;
+    NSMutableArray *paths = [url.pathComponents mutableCopy];
+    // [0] = /
+    // [1] = projectId
+    // If there are more than two, that means someone wants to have an additional path
+    // ref: https://github.com/getsentry/sentry-cocoa/issues/236
+    NSString *path = @"";
+    if ([paths count] > 2) {
+        [paths removeObjectAtIndex:0]; // We remove the leading /
+        [paths removeLastObject]; // We remove projectId since we add it later
+        path = [NSString stringWithFormat:@"/%@", [paths componentsJoinedByString:@"/"]]; // We put together the path
+    }
+    NSURLComponents *components = [NSURLComponents new];
+    components.scheme = url.scheme;
+    components.host = url.host;
+    components.port = url.port;
+    components.path = [NSString stringWithFormat:@"%@/api/%@/", path, projectId];
+    return components.URL;
 }
 
 - (NSURL *_Nullable)convertDsnString:(NSString *)dsnString didFailWithError:(NSError *_Nullable *_Nullable)error {