+++ Sources/Sentry/include/SentryDsn.h
@@ -4,11 +4,14 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface SentryDsn : NSObject
 
-@property(nonatomic, strong) NSURL *url;
+@property(nonatomic, strong, readonly) NSURL *url;
 
 - (_Nullable instancetype)initWithString:(NSString *)dsnString didFailWithError:(NSError *_Nullable *_Nullable)error;
 
 - (NSString *)getHash;
+
+- (NSURL *)getStoreEndpoint;
+- (NSURL *)getEnvelopeEndpoint;
 
 @end