+++ Tests/SentryTests/Networking/SentryDsnTests.m
@@ -11,17 +11,9 @@
 #import "SentryError.h"
 #import "SentryDsn.h"
 
-@interface SentryNSURLRequest (Private)
-
-+ (NSURL *)getStoreUrlFromDsn:(SentryDsn *)dsn;
-
-@end
-
 @interface SentryDsnTests : XCTestCase
 
 @end
-
-//+ (NSURL *)getStoreUrlFromDsn:(SentryDsn *)dsn
 
 @implementation SentryDsnTests
 
@@ -80,18 +72,47 @@
     XCTAssertEqual(kSentryErrorInvalidDsnError, error.code);
     XCTAssertNil(options);
 }
-    
+
 - (void)testDsnUrl {
     NSError *error = nil;
     SentryDsn *dsn = [[SentryDsn alloc] initWithString:@"https://username:password@getsentry.net/1" didFailWithError:&error];
-    
-    XCTAssertEqualObjects([[SentryNSURLRequest getStoreUrlFromDsn:dsn] absoluteString], @"https://getsentry.net/api/1/store/");
+
+    XCTAssertEqualObjects([[dsn getStoreEndpoint] absoluteString], @"https://getsentry.net/api/1/store/");
     XCTAssertNil(error);
-    
+
     SentryDsn *dsn2 = [[SentryDsn alloc] initWithString:@"https://username:password@sentry.io/foo/bar/baz/1" didFailWithError:&error];
-    
-    XCTAssertEqualObjects([[SentryNSURLRequest getStoreUrlFromDsn:dsn2] absoluteString], @"https://sentry.io/foo/bar/baz/api/1/store/");
+
+    XCTAssertEqualObjects([[dsn2 getStoreEndpoint] absoluteString], @"https://sentry.io/foo/bar/baz/api/1/store/");
     XCTAssertNil(error);
+}
+
+- (void)testGetEnvelopeUrl {
+    NSError *error = nil;
+    SentryDsn *dsn = [[SentryDsn alloc] initWithString:@"https://username:password@getsentry.net/1" didFailWithError:&error];
+
+    XCTAssertEqualObjects([[dsn getEnvelopeEndpoint] absoluteString], @"https://getsentry.net/api/1/envelope/");
+    XCTAssertNil(error);
+
+    SentryDsn *dsn2 = [[SentryDsn alloc] initWithString:@"https://username:password@sentry.io/foo/bar/baz/1" didFailWithError:&error];
+
+    XCTAssertEqualObjects([[dsn2 getEnvelopeEndpoint] absoluteString], @"https://sentry.io/foo/bar/baz/api/1/envelope/");
+    XCTAssertNil(error);
+}
+
+- (void)testGetStoreDsnCachesResult {
+    SentryDsn *dsn = [[SentryDsn alloc] initWithString:@"https://username:password@getsentry.net/1" didFailWithError:nil];
+
+    XCTAssertNotNil([dsn getStoreEndpoint]);
+    // Assert same reference
+    XCTAssertTrue([dsn getStoreEndpoint] == [dsn getStoreEndpoint]);
+}
+
+- (void)testGetEnvelopeDsnCachesResult {
+    SentryDsn *dsn = [[SentryDsn alloc] initWithString:@"https://username:password@getsentry.net/1" didFailWithError:nil];
+
+    XCTAssertNotNil([dsn getEnvelopeEndpoint]);
+    // Assert same reference
+    XCTAssertTrue([dsn getEnvelopeEndpoint] == [dsn getEnvelopeEndpoint]);
 }
 
 @end