+++ Tests/SentryTests/SentryScopeTests.m
@@ -115,13 +115,6 @@
     #warning TODO implement
 }
 
-- (void)testReleaseSerializes {
-    SentryScope *scope = [[SentryScope alloc] init];
-    NSString *expectedReleaseName = @"io.sentry.cocoa@5.0.0-deadbeef";
-    [scope setRelease:expectedReleaseName];
-    XCTAssertEqualObjects([[scope serialize] objectForKey:@"release"], expectedReleaseName);
-}
-
 - (void)testDistSerializes {
     SentryScope *scope = [[SentryScope alloc] init];
     NSString *expectedDist = @"dist-1.0";
@@ -162,7 +155,6 @@
     [scope addBreadcrumb:[self getBreadcrumb]];
     [scope setUser:[[SentryUser alloc] initWithUserId:@"id"]];
     [scope setContextValue:@{@"e": @"f"} forKey:@"myContext"];
-    [scope setRelease:@"123"];
     [scope setDist:@"456"];
     [scope setEnvironment:@"789"];
     [scope setFingerprint:@[@"a"]];
@@ -177,7 +169,6 @@
     [cloned addBreadcrumb:[[SentryBreadcrumb alloc] initWithLevel:kSentryLevelDebug category:@"http2"]];
     [cloned setUser:[[SentryUser alloc] initWithUserId:@"aid"]];
     [cloned setContextValue:@{@"ae": @"af"} forKey:@"myContext"];
-    [cloned setRelease:@"a123"];
     [cloned setDist:@"a456"];
     [cloned setEnvironment:@"a789"];