+++ CHANGELOG.md
@@ -2,9 +2,12 @@
 
 ## unreleased
 
+- fix: Use RateLimitCategoryError for events #470
 - feat: Store SentryEnvelopes in extra path #468
 - feat: Adds setUser to SentrySDK and SentryHub #467
 - feat: Add auto session starting for macOS #463
+- fix: Take release name from options #462
+- feat: Use new envelope endpoint #475
 
 ## 5.0.0-beta.6